Fix resetting the session for non-html requests
this caused a CSRF issue on saas. Requests were being redirected to
/admin/subscriptions/expired and wrongfully so for API and CSS requests.

Now imagine a request to `/` with `login_required` for a non-logged in
user

1. The GET `/` request is redirected to `/login?back_url=...` by
ApplicationController. Session is reset here to X

2. The user receives the login form with CSRF value from session X

3. A request is made to `/api/v3/configuration` from the frontend. It
too is redirected to a Rails controller and ends up in the same
`check_if_login_required` flow, resetting the session again

4. The user now has a session cookie Y, but the CSRF form is still from
X

5. submitting it results in 422 CSRF validation error